requirements.txt
django
djangorestframework
gunicorn
coverage
selenium
locust